requirements.txt
argcomplete==1.8.2
base58==0.2.4
boto3==1.4.4
botocore==1.5.40
certifi==2018.4.16
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
Django==2.0.7
django-filter==2.0.0
django-filters==0.2.1
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
durationpy==0.4
future==0.16.0
futures==3.1.1
hjson==2.0.2
idna==2.7
jmespath==0.9.2
jsonfield==2.0.2
kappa==0.6.0
lambda-packages==0.15.0
munch==2.3.2
mysqlclient==1.3.13
numpy==1.15.4
Pillow==3.1.1
placebo==0.8.1
psycopg2==2.7.6.1
python-dateutil==2.6.0
python-slugify==1.2.4
pytz==2018.5
PyYAML==3.12
qanda-module==1.0.0
qbase-module==4.4.1
requests==2.19.1
s3transfer==0.1.13
Serializer==0.2.1
serializers==0.2.4
simplejson==3.16.0
six==1.10.0
toml==0.9.2
tqdm==4.11.2
troposphere==2.3.1
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.42.0
